Refacter controller commit function to models
